+++ node/ts/net.ts
@@ -688,6 +688,18 @@ export type NetConstructorOptions = Readonly<
     }
 >;
 
+/** See {@link Net.setProxy()}. */
+export type ProxyOptions = Readonly<{
+  scheme: string;
+  host: string;
+  port?: number;
+  username?: string;
+  password?: string;
+}>;
+
+/** The "scheme" for Signal TLS proxies. See {@link Net.setProxy()}. */
+export const SIGNAL_TLS_PROXY_SCHEME = 'org.signal.tls';
+
 export class Net {
   private readonly asyncContext: TokioAsyncContext;
   private readonly connectionManager: ConnectionManager;
@@ -825,14 +837,51 @@ export class Net {
   /**
    * Sets the proxy host to be used for all new connections (until overridden).
    *
-   * Sets a domain name and port to be used to proxy all new outgoing
-   * connections. The proxy can be overridden by calling this method again or
-   * unset by calling {@link #clearProxy}.
+   * Sets a server to be used to proxy all new outgoing connections. The proxy can be overridden by
+   * calling this method again or unset by calling {@link #clearProxy}. Omitting the `port` means
+   * the default port for the scheme will be used.
+   *
+   * To specify a Signal transparent TLS proxy, use {@link SIGNAL_TLS_PROXY_SCHEME}, or the
+   * overload that takes a separate domain and port number.
+   *
+   * Throws if the scheme is unsupported or if the provided parameters are invalid for that scheme
+   * (e.g. Signal TLS proxies don't support authentication)
+   */
+  setProxy(options: ProxyOptions): void;
+  /**
+   * Sets the Signal TLS proxy host to be used for all new connections (until overridden).
+   *
+   * Sets a domain name and port to be used to proxy all new outgoing connections, using a Signal
+   * transparent TLS proxy. The proxy can be overridden by calling this method again or unset by
+   * calling {@link #clearProxy}.
    *
    * Throws if the host or port is structurally invalid, such as a port that doesn't fit in u16.
    */
-  setProxy(host: string, port: number): void {
-    Native.ConnectionManager_set_proxy(this.connectionManager, host, port);
+  setProxy(host: string, port?: number): void;
+  setProxy(hostOrOptions: string | ProxyOptions, portOrNothing?: number): void {
+    if (typeof hostOrOptions === 'string') {
+      // Support <username>@<host> syntax to allow UNENCRYPTED_FOR_TESTING as a marker user.
+      // This is not a stable feature of the API and may go away in the future;
+      // the Rust layer will reject any other users anyway. But it's convenient for us.
+      const [before, after] = hostOrOptions.split('@', 2);
+      const [username, domain] = after ? [before, after] : [undefined, before];
+      hostOrOptions = {
+        scheme: SIGNAL_TLS_PROXY_SCHEME,
+        host: domain,
+        port: portOrNothing,
+        username,
+      };
+    }
+    const { scheme, host, port, username, password } = hostOrOptions;
+    Native.ConnectionManager_set_proxy(
+      this.connectionManager,
+      scheme,
+      host,
+      // i32::MIN represents "no port provided"; we don't expect anyone to pass that manually.
+      port ?? -0x8000_0000,
+      username ?? null,
+      password ?? null
+    );
   }
 
   /**