+++ node/ts/test/NetTest.ts
@@ -21,6 +21,7 @@ import {
   Environment,
   Net,
   newNativeHandle,
+  SIGNAL_TLS_PROXY_SCHEME,
 } from '../net';
 import { ChatResponse } from '../../Native';
 import { CompletablePromise } from './util';
@@ -169,7 +170,6 @@ describe('chat service api', () => {
   });
 
   it('invalid proxies are rejected', () => {
-    // The default TLS proxy config doesn't support staging, so we connect to production.
     const net = new Net({
       env: Environment.Production,
       userAgent: userAgent,
@@ -178,6 +178,9 @@ describe('chat service api', () => {
     expect(() => net.setProxy('signalfoundation.org', 100_000)).throws(Error);
     expect(() => net.setProxy('signalfoundation.org', -1)).throws(Error);
     expect(() => net.setProxy('signalfoundation.org', 0.1)).throws(Error);
+    expect(() =>
+      net.setProxy({ scheme: 'socks+shoes', host: 'signalfoundation.org' })
+    ).throws(Error);
   });
 
   // Integration tests make real network calls and as such will not be run unless a proxy server is provided.
@@ -248,6 +251,35 @@ describe('chat service api', () => {
             2
           );
           net.setProxy(host, parseInt(port, 10));
+          await connectChatUnauthenticated(net);
+        }).timeout(10000);
+
+        it('can connect through a proxy server using the options API', async function () {
+          const PROXY_SERVER = process.env.LIBSIGNAL_TESTING_PROXY_SERVER;
+          if (!PROXY_SERVER) {
+            this.skip();
+          }
+
+          // The default TLS proxy config doesn't support staging, so we connect to production.
+          const net = new Net({
+            env: Environment.Production,
+            userAgent: userAgent,
+          });
+          const [host = PROXY_SERVER, port = '443'] = PROXY_SERVER.split(
+            ':',
+            2
+          );
+          const [before, after] = host.split('@', 2);
+          const [username, domain] = after
+            ? [before, after]
+            : [undefined, before];
+
+          net.setProxy({
+            scheme: SIGNAL_TLS_PROXY_SCHEME,
+            host: domain,
+            port: parseInt(port, 10),
+            username,
+          });
           await connectChatUnauthenticated(net);
         }).timeout(10000);
       });